kmk: warning
svn:sync-xref-src-repo-rev: r3643

+++ src/kmk/strcache2.h
@@ -1,4 +1,4 @@
-/* $Id: strcache2.h 2413 2010-09-11 17:43:04Z knut.osmundsen@oracle.com $ */
+/* $Id: strcache2.h 3643 2024-11-02 23:40:44Z knut.osmundsen@oracle.com $ */
 /** @file
  * strcache - New string cache.
  */
@@ -157,7 +157,7 @@ MY_INLINE unsigned int
 strcache2_calc_ptr_hash (struct strcache2 *cache, const char *str)
 {
   (void)cache;
-  return (size_t)str >> STRCACHE2_ENTRY_ALIGN_SHIFT;
+  return (unsigned int) ((size_t)str >> STRCACHE2_ENTRY_ALIGN_SHIFT);
 }
 
 /* Get the user value for the string. */
